# pkg/datasource/expr: improve performance