@flyth 2025-06-23
> Thanks.
> 
> This PR does not expose the feature on the CLI yet, right?

No, this is just speeding up the internals for now.